updated test for config_store
