+++ tests/conftest.py
@@ -1,7 +1,7 @@
 import os
 import pytest
 from user_sync import config
-
+import shutil
 
 @pytest.fixture
 def fixture_dir():
@@ -25,3 +25,29 @@ def cli_args():
             args_out[k] = v
         return args_out
     return _cli_args
+
+
+@pytest.fixture
+def tmp_config_files(root_config_file, ldap_config_file, umapi_config_file, tmpdir):
+    tmpfiles = []
+    for fname in [root_config_file, ldap_config_file, umapi_config_file]:
+        basename = os.path.split(fname)[-1]
+        tmpfile = os.path.join(str(tmpdir), basename)
+        shutil.copy(fname, tmpfile)
+        tmpfiles.append(tmpfile)
+    return tuple(tmpfiles)
+
+@pytest.fixture
+def root_config_file(fixture_dir):
+    return os.path.join(fixture_dir, 'user-sync-config.yml')
+
+
+@pytest.fixture
+def ldap_config_file(fixture_dir):
+    return os.path.join(fixture_dir, 'connector-ldap.yml')
+
+
+@pytest.fixture
+def umapi_config_file(fixture_dir):
+    return os.path.join(fixture_dir, 'connector-umapi.yml')
+
